+++ TabLayout/TabLayout.js
@@ -189,8 +189,6 @@ const TabLayoutBase = kind({
 		/**
 		 * Orientation of the tabs.
 		 *
-		 * Horizontal tabs support a maximum of six tabs.
-		 *
 		 * @type {('horizontal'|'vertical')}
 		 * @default 'vertical'
 		 * @public
@@ -336,14 +334,13 @@ const TabLayoutBase = kind({
 			'--tablayout-expand-collapse-diff': ((orientation === 'vertical') ? scaleToRem(dimensions.tabs.normal - dimensions.tabs.collapsed) : 0)
 		}),
 		tabOrientation: ({orientation}) => orientation === 'vertical' ? 'horizontal' : 'vertical',
-		// limit to 6 tabs for horizontal orientation
-		tabs: ({children, orientation}) => {
+		tabs: ({children}) => {
 			const tabs = mapAndFilterChildren(children, (child) => (
 				Object.keys(child.props)
 					.filter((prop) => prop !== 'children' && prop !== 'id')
 					.reduce((obj, key) => ({...obj, [key]: child.props[key]}), {})
 			));
-			return orientation === 'horizontal' && tabs.length > 6 ? tabs.slice(0, 6) : tabs;
+			return tabs;
 		}
 	},