+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `sandstone/TabLayout` to eliminate the horizontal maximum number of tabs
+
 ## [2.0.9] - 2022-05-19
 
 ### Fixed